Center CSV Table Results and Fix Legend UI Redundancy
Centered the CSV results table beneath the graph and legend for a cleaner UI.
Removed the redundancy of a legend in LineGraph.tsx.
Fixed buggy "Select All" feature which had an issue where reselecting "Select All" caused the legend and table to disappear.
Formatted using Prettier for readability and added comments for further UI/UX changes.